Move research story into-assistant testing
