+++ packages/core/assistant-testing/package.json
@@ -32,6 +32,8 @@
     "@dxos/context": "workspace:*",
     "@dxos/echo": "workspace:*",
     "@dxos/echo-db": "workspace:*",
+    "@dxos/echo-protocol": "workspace:*",
+    "@dxos/echo-schema": "workspace:*",
     "@dxos/edge-client": "workspace:*",
     "@dxos/effect": "workspace:*",
     "@dxos/functions": "workspace:*",
@@ -41,7 +43,8 @@
     "@dxos/plugin-markdown": "workspace:*",
     "@dxos/react-ui-syntax-highlighter": "workspace:*",
     "@dxos/schema": "workspace:*",
-    "@dxos/util": "workspace:*"
+    "@dxos/util": "workspace:*",
+    "exa-js": "^1.5.12"
   },
   "devDependencies": {
     "@effect/ai": "0.23.0",
