+++ packages/core/assistant-testing/src/functions/research/instructions-research.tpl
