+++ packages/core/assistant-testing/tsconfig.json
@@ -60,6 +60,12 @@
       "path": "../echo/echo-db"
     },
     {
+      "path": "../echo/echo-protocol"
+    },
+    {
+      "path": "../echo/echo-schema"
+    },
+    {
       "path": "../functions"
     },
     {
