+++ packages/core/assistant/package.json
@@ -68,7 +68,6 @@
     "@preact/signals-core": "^1.9.0",
     "@xenova/transformers": "^2.17.2",
     "chalk": "^4.1.0",
-    "exa-js": "^1.5.12",
     "handlebars": "4.7.8",
     "lodash.defaultsdeep": "^4.6.1"
   },
