+++ packages/core/assistant/src/index.ts
@@ -4,7 +4,6 @@
 
 export * from './conversation';
 export * from './functions';
-export * from './research';
 export * from './session';
 export * from './templates';
 export * from './types';
